+++ src/components/Catalogs.vue
@@ -6,24 +6,38 @@
       <ViewButtons v-if="!hideControls" class="me-2" v-model="view" />
       <SortButtons v-if="!hideControls && isComplete && catalogs.length > 1" v-model="sort.direction" />
     </header>
-    <section v-if="!hideControls && isComplete && catalogs.length > 1" class="catalog-filter mb-2">
-      <SearchBox v-model="searchTerm" :placeholder="filterPlaceholder" />
-      <multiselect
-        v-if="allKeywords.length > 0"
-        v-model="selectedKeywords"
-        :options="allKeywords"
-        multiple
-        :placeholder="$t('multiselect.keywordsPlaceholder')"
-        :select-label="$t('multiselect.selectLabel')"
-        :selected-label="$t('multiselect.selectedLabel')"
-        :deselect-label="$t('multiselect.deselectLabel')"
-        :limit-text="limitText"
-      />
+    <section v-if="!hideControls && ((isComplete && catalogs.length > 1) || canSearchFreeText)" class="catalog-filter mb-2">
+      <template v-if="canSearchFreeText">
+        <multiselect
+          multiple taggable @tag="addSearchTerm"
+          id="catalogFreeText" v-model="selectedSearchTerms" :options="selectedSearchTerms"
+          :placeholder="$t('search.enterSearchTerms')" :tag-placeholder="$t('search.addSearchTerm')" :no-options="$t('search.addSearchTerm')"
+        >
+          <template #noOptions>{{ $t('search.noOptions') }}</template>
+        </multiselect>
+        <b-button v-if="advancedSearchLink" variant="primary" :to="advancedSearchLink" class="additional-filter-link">
+          {{ $t('search.additionalFilters') }}
+        </b-button>
+      </template>
+      <template v-else>
+        <SearchBox v-model="searchTerm" :placeholder="filterPlaceholder" />
+        <multiselect
+          v-if="isComplete && allKeywords.length > 0"
+          v-model="selectedKeywords"
+          :options="allKeywords"
+          multiple
+          :placeholder="$t('multiselect.keywordsPlaceholder')"
+          :select-label="$t('multiselect.selectLabel')"
+          :selected-label="$t('multiselect.selectedLabel')"
+          :deselect-label="$t('multiselect.deselectLabel')"
+          :limit-text="limitText"
+        />
+      </template>
     </section>
     <Pagination v-if="showPagination" ref="topPagination" class="mb-3" :pagination="pagination" placement="top" @paginate="paginate" />
     <b-alert v-if="hasSearchCritera && catalogView.length === 0" variant="warning" class="mt-2" show>{{ $t('catalogs.noMatches') }}</b-alert>
     <section class="list">
-      <Loading v-if="loading" fill top />
+      <Loading v-if="loading && !loadingMore" fill top />
       <div :class="view === 'list' ? 'card-list' : 'card-grid'">
         <Catalog v-for="catalog in catalogView" :catalog="catalog" :key="catalog.href">
           <template #footer="{data}">
@@ -33,7 +47,10 @@
       </div>
     </section>
     <Pagination v-if="showPagination" class="mb-3" :pagination="pagination" @paginate="paginate" />
-    <b-button v-else-if="hasMore" @click="loadMore" variant="primary" v-visible.300="loadMore">{{ $t('catalogs.loadMore') }}</b-button>
+    <b-button v-else-if="hasMore" @click="loadMore" variant="primary" v-visible.300="loadMore">
+      <b-spinner v-if="loading && loadingMore" small />
+      {{ $t('catalogs.loadMore') }}
+    </b-button>
   </section>
 </template>
 
@@ -47,6 +64,8 @@ import { STAC } from 'stac-js';
 import ViewButtons from './ViewButtons.vue';
 import Utils from '../utils';
 import { sortStac } from '../models/stac';
+import { TYPES } from './ApiCapabilitiesMixin.js';
+import { hasText, URI } from 'stac-js/src/utils.js';
 
 export default defineComponent({
   name: "Catalogs",
@@ -80,6 +99,10 @@ export default defineComponent({
       type: Boolean,
       default: false
     },
+    loadingMore: {
+      type: Boolean,
+      default: false
+    },
     hasMore: {
       type: Boolean,
       default: false
@@ -87,6 +110,10 @@ export default defineComponent({
     apiFilters: {
       type: Object,
       default: () => ({})
+    },
+    apiSearch: {
+      type: Boolean,
+      default: false
     },
     pagination: {
       type: Object,
@@ -97,18 +124,32 @@ export default defineComponent({
       default: null
     }
   },
-  emits: ['loadMore', 'paginate'],
+  emits: ['loadMore', 'paginate', 'search'],
   data() {
     return {
       searchTerm: '',
       sort: Utils.parseApiSortParameter(), // get empty sort object
-      selectedKeywords: []
+      selectedKeywords: [],
+      selectedSearchTerms: [],
     };
   },
   computed: {
     ...mapState(['defaultCollectionSort', 'uiLanguage']),
-    ...mapGetters(['supportsConformance']),
+    ...mapGetters(['searchBrowserLink', 'supportsConformance']),
     ...mapGetters(['getStac']),
+    advancedSearchLink() {
+      if (!this.canSearchFreeText || !this.searchBrowserLink) {
+        return null;
+      }
+      const query = Utils.stateQueryParametersToObject({
+        'searchtype': 'collections',
+        'q': this.selectedSearchTerms,
+      });
+      return URI(this.searchBrowserLink).query(query).toString();
+    },
+    canSearchFreeText() {
+      return this.apiSearch && this.supportsConformance(TYPES.Collections.FreeText);
+    },
     catalogCount() {
       if (this.catalogs.length !== this.catalogView.length) {
         return this.catalogView.length + '/' + this.catalogs.length;
@@ -140,7 +181,7 @@ export default defineComponent({
     },
     showPagination() {
       // Check whether any pagination links are available
-      return Object.values(this.pagination).some(link => !!link);
+      return Object.values(this.pagination).some(link => Boolean(link));
     },
     allCatalogs() {
       return this.catalogs.map(catalog => {
@@ -149,7 +190,7 @@ export default defineComponent({
       });
     },
     hasSearchCritera() {
-      return this.searchTerm || this.selectedKeywords.length > 0;
+      return this.searchTerm || this.selectedKeywords.length > 0 || this.selectedSearchTerms.length > 0;
     },
     catalogView() {
       if (this.hasMore) {
@@ -218,10 +259,24 @@ export default defineComponent({
       }
     }
   },
+  watch: {
+    selectedSearchTerms: {
+      handler(searchTerms) {
+        this.$emit('search', searchTerms);
+      },
+      deep: 1
+    }
+  },
   created() {
     this.sort = Utils.parseApiSortParameter(this.defaultCollectionSort);
   },
   methods: {
+    addSearchTerm(term) {
+      if (!hasText(term)) {
+        return;
+      }
+      this.selectedSearchTerms.push(term);
+    },
     loadMore(visible = true) {
       if (visible) {
         this.$emit('loadMore');
@@ -253,6 +308,12 @@ export default defineComponent({
     flex-grow: 1;
     flex-basis: 300px;
     min-width: 300px;
+  }
+
+  > .additional-filter-link {
+    flex-grow: 0;
+    align-self: center;
+    white-space: nowrap;
   }
 }
 </style>