+++ src/views/Catalog.vue
@@ -48,7 +48,7 @@
         </section>
         <section v-if="isCollection || hasThumbnails" class="mb-4">
           <b-card no-body class="maps-preview">
-            <b-tabs v-model="tab" ref="tabs" pills card vertical end>
+            <b-tabs v-model="tab" pills card vertical end>
               <b-tab v-if="isCollection" :id="tabIds.map" :title="$t('map')" no-body>
                 <MapView :stac="data" v-bind="mapData" @changed="dataChanged" @empty="handleEmptyMap" onfocusOnly popover />
               </b-tab>
@@ -67,13 +67,17 @@
       </b-col>
       <b-col class="catalogs-container" v-if="hasCatalogs">
         <WidgetHook id="view-catalog-catalogs-start" />
-        <Catalogs :catalogs="catalogs" :hasMore="hasMore" @load-more="loadMoreCollections" />
+        <Catalogs
+          :apiSearch="hasApiCollections" :catalogs="catalogs" :hasMore="hasMore"
+          @load-more="loadMoreCollections" @search="searchCollections"
+          :loading="Boolean(loadingCollections)" :loadingMore="loadingCollections === 'more'"
+        />
         <WidgetHook id="view-catalog-catalogs-end" />
       </b-col>
       <b-col class="items-container" v-if="hasItems || hasItemAssets">
         <WidgetHook id="view-catalog-items-start" />
         <Items
-          :stac="data" :items="items" :api="isApi"
+          :stac="data" :items="items" :api="hasApiItems"
           :showFilters="showFilters" :apiFilters="filters"
           :pagination="itemPages" :loading="apiItemsLoading"
           :count="apiItemsNumberMatched"
@@ -98,7 +102,7 @@ import ShowAssetLinkMixin from '../components/ShowAssetLinkMixin';
 import StacFieldsMixin from '../components/StacFieldsMixin';
 import { formatLicense, formatTemporalExtents } from '@radiantearth/stac-fields/formatters';
 import Utils from '../utils';
-import { hasText, isObject } from 'stac-js/src/utils.js';
+import { hasText, isObject, size } from 'stac-js/src/utils.js';
 import { addSchemaToDocument, createCatalogSchema } from '../schema-org';
 import { ItemCollection } from '../models/stac.js';
 import DeprecationMixin from '../components/DeprecationMixin.js';
@@ -137,11 +141,14 @@ export default defineComponent({
   ],
   data() {
     return {
-      filters: {}
+      filters: {},
+      loadingCollections: null,
+      isSearchingCollections: false,
+      currentSearchRequestId: 0,
     };
   },
   computed: {
-    ...mapState(['data', 'url', 'apiCatalogPriority',  'apiItems', 'apiItemsLink', 'apiItemsPagination', 'apiItemsNumberMatched', 'nextCollectionsLink', 'stateQueryParameters']),
+    ...mapState(['data', 'apiCatalogPriority', 'apiItemsLink', 'apiItemsPagination', 'apiItemsNumberMatched', 'nextCollectionsLink', 'stateQueryParameters']),
     ...mapGetters(['catalogs', 'collectionLink', 'isCollection', 'items', 'getApiItemsLoading', 'parentLink', 'rootLink']),
     droppedCql() {
       return this.$store.state.search.droppedFilters.filter(f => f.type === 'cql2');
@@ -163,7 +170,7 @@ export default defineComponent({
       return null;
     },
     showFilters() {
-      return Boolean(this.stateQueryParameters['itemFilterOpen']);
+      return Boolean(this.stateQueryParameters.itemFilterOpen);
     },
     linkPosition() {
       if (this.additionalLinks.length === 0) {
@@ -226,14 +233,17 @@ export default defineComponent({
       }
       return pages;
     },
-    isApi() {
+    hasApiItems() {
       return Boolean(this.apiItemsLink);
     },
+    hasApiCollections() {
+      return Boolean(this.data.getApiCollectionsLink()) && this.apiCatalogPriority !== 'childs';
+    },
     hasItems() {
-      return this.items.length > 0 || this.isApi;
+      return this.items.length > 0 || this.hasApiItems;
     },
     hasCatalogs() {
-      return this.catalogs.length > 0;
+      return this.catalogs.length > 0 || this.hasApiCollections || this.isSearchingCollections;
     },
     mapData() {
       const data = {};
@@ -295,8 +305,50 @@ export default defineComponent({
     filtersShown(show) {
       this.$store.commit('updateState', {type: 'itemFilterOpen', value: show ? 1 : null});
     },
-    loadMoreCollections() {
-      this.$store.dispatch('loadNextApiCollections', {show: true});
+    async loadMoreCollections() {
+      const requestId = this.currentSearchRequestId;
+      this.loadingCollections = "more";
+      try {
+        const params = {
+          show: true,
+          searching: this.isSearchingCollections
+        };
+        if (this.isSearchingCollections) {
+          params.searchRequestId = requestId;
+        }
+        await this.$store.dispatch('loadNextApiCollections', {
+          ...params
+        });
+      } catch (error) {
+        this.$store.commit('showGlobalError', {
+          error,
+          message: this.$t('errors.loadApiCollectionsFailed')
+        });
+      } finally {
+        if (requestId === this.currentSearchRequestId && this.loadingCollections === 'more') {
+          this.loadingCollections = null;
+        }
+      }
+    },
+    async searchCollections(searchTerms) {
+      this.loadingCollections = "all";
+      this.isSearchingCollections = size(searchTerms) > 0;
+      // Increment request ID to invalidate any in-flight requests from previous searches
+      const requestId = ++this.currentSearchRequestId;
+      try {
+        await this.$store.dispatch('loadNextApiCollections', {
+          stac: this.data, show: true, q: searchTerms, searching: this.isSearchingCollections, searchRequestId: requestId
+        });
+      } catch (error) {
+        this.$store.commit('showGlobalError', {
+          error,
+          message: this.$t('errors.loadApiCollectionsFailed')
+        });
+      } finally {
+        if (requestId === this.currentSearchRequestId && this.loadingCollections === 'all') {
+          this.loadingCollections = null;
+        }
+      }
     },
     async paginateItems(link) {
       try {